Refactors cli
